Payment v5

+++ DAL/MenuItemDao.cs
@@ -45,7 +45,6 @@ namespace DAL
         }
 
 
-        // This method is used to convert the DataTable to a List of Menu_Item_Model
         private List<MenuItemModel> ConvertToList(DataTable dataTable)
         {
             List<MenuItemModel> menuItemList = new List<MenuItemModel>();
@@ -57,7 +56,7 @@ namespace DAL
                     Name = row["name"].ToString(),
                     Item_Category = row["item_category"].ToString(),
                     Stock = Convert.ToInt32(row["stock"]),
-                    Vat = Convert.ToDouble(row["vat"]),
+                    Vat = Convert.ToDecimal(row["vat"]),
                     Price = Convert.ToDecimal(row["price"]),
                     Preperation_Time = Convert.ToInt32(row["preparation_time"]),
                     Menu_Id = row.Table.Columns.Contains("menu_id") && row["menu_id"] != DBNull.Value // in case if its null.
