+++ Model/Bill.cs
@@ -10,7 +10,7 @@ namespace Model
     {
         public int BillId { get; set; }
         public float TotalPrice { get; set; }
-        public float Vat { get; set; }
+        public decimal Vat { get; set; }
         public int GuestNumber { get; set; }
         public int OrderId { get; set; }
         public float Tip { get; set; }
